+++ src/validation/SourceFile.h
@@ -24,7 +24,7 @@ class SourceFile
      * @param line The line number, starting from 1.
      * @return A view of the line.
      */
-    [[nodiscard]] std::string_view getLine(unsigned int line) const;
+    [[nodiscard]] std::string_view getLine(size_t line) const;
 
     /**
      * Get a string view of a slice of a line.
@@ -33,12 +33,11 @@ class SourceFile
      * @param column_end The last column to include.
      * @return A view of the slice.
      */
-    [[nodiscard]] std::string_view getLineSlice(unsigned int line,
-                                                unsigned int column_start,
-                                                unsigned int column_end) const;
+    [[nodiscard]] std::string_view getLineSlice(size_t line, unsigned int column_start, unsigned int column_end) const;
 
   private:
     std::vector<std::string> lines_;
 };
 
 #endif
+
